requirements.txt
ibm_db==3.1.1
psutil==5.9.0

aiofiles~=24.1.0
PyYAML~=6.0.2

